Created new Runnit file and moved some files over for testing
