+++ robot/robot.py
@@ -1,15 +1,24 @@
+import timer
 import wpilib
 import rev
 from ctre import WPI_TalonSRX
 from magicbot import MagicRobot
 from networktables import NetworkTables, NetworkTable
 from wpilib import DoubleSolenoid
-from components.drivetrain import DriveTrain
+import wpilib.drive
+from robotpy_ext.autonomous import AutonomousModeSelector
 
-import time
-from components.boom import Boom
-from components.grabber import grabber
 
+from wpilib import MotorControllerGroup
+from wpilib.drive import DifferentialDrive
+from networktables import NetworkTable
+from ctre import WPI_TalonSRX
+# -------------------------------- Commented Out Imports, to be un-commented ---------------------
+# from robot.utils import limit
+# from components.drivetrain import DriveTrain
+# from components.boom import Boom
+# from components.grabber import grabber
+# ------------------------------------------------------------------------------------------------
 # Download and install stuff on the RoboRIO after imaging
 '''py -3 -m robotpy_installer download-python
    py -3 -m robotpy_installer install-python
@@ -32,11 +41,76 @@ from components.grabber import grabber
 INPUT_SENSITIVITY = .3
 
 MagicRobot.control_loop_wait_time = 0.05
+
+SPEED_MULTIPLIER = 1
+ANGLE_MULTIPLIER = 1
+
+# ---------------------------------------------------------------------------------------------------------------------
+# Section added from drivetrain.py for testing autonomous, will be removed after completion
+def limit(number: float, limits: list) -> float:
+    """
+    return a number within the limits, returning the limiting number if out of bounds
+    number: any number
+    limits: list of limits, first element is minimum, second is max
+    ex:
+        >>> limit(5, [-4, 4])
+        >>> 4
+    """
+    return min(max(number, limits[0]), limits[1])
+class DriveTrain:
+    # Magicrobot variable injection automatically gets these values from main robot
+    # NOTE: components must go in the components folder for magicrobot to properly inject shared variables
+    talon_L_1: WPI_TalonSRX
+    talon_L_2: WPI_TalonSRX
+
+    talon_R_1: WPI_TalonSRX
+    talon_R_2: WPI_TalonSRX
+
+    sd: NetworkTable
+
+
+
+    def setup(self):
+        """instead of __init__(), use setup() to initialize values (works with magicirobot variable injection"""
+
+        self.left_motors: MotorControllerGroup = MotorControllerGroup(self.talon_L_1, self.talon_L_2)
+        self.right_motors: MotorControllerGroup = MotorControllerGroup(self.talon_R_1, self.talon_R_2)
+
+        self.drive: DifferentialDrive = DifferentialDrive(self.left_motors, self.right_motors)
+
+    # control method
+    def set_motors(self, speed: float, angle: float):
+        """
+        sets the speed and angle of the motors
+        speed: percentage of full speed
+        angle: percentage of full rotation, ccw is positive
+        Puts values into smartdashboard to be called by arcadeDrive() later in execute()
+        """
+
+        self.sd.putValue("Speed", limit(speed, [-1, 1]))
+        self.sd.putValue("Angle", limit(angle, [-1, 1]))
+
+    def execute(self) -> None:
+        """
+        Reads the data from smartdashboard (set by control methods), and then sends data to output devices such as motors.
+        Execute is called in telopPeriodic automatically; no need to manually call
+        """
+
+        speed = self.sd.getValue("Speed", defaultValue=0.0)
+        angle = self.sd.getValue("Angle", defaultValue=0.0)
+
+        # print(speed, angle)
+
+        self.drive.arcadeDrive(angle * ANGLE_MULTIPLIER, speed * SPEED_MULTIPLIER, True)
+        # NOTE: THIS IS INVERSED?
+        # self.drive.arcadeDrive(speed * SPEED_MULTIPLIER, angle * ANGLE_MULTIPLIER, squareInputs=True)
+
+# -------------------------------------------------------------------------------------------------------------------------
 class SpartaBot(MagicRobot):
 
     # a DriveTrain instance is automatically created by MagicRobot
     drivetrain: DriveTrain
-    boom_arm: Boom
+    # boom_arm: Boom
 
     def createObjects(self):
         '''Create motors and stuff here'''
@@ -46,6 +120,9 @@ class SpartaBot(MagicRobot):
 
         NetworkTables.initialize(server='roborio-5045-frc.local')
         self.sd: NetworkTable = NetworkTables.getTable('SmartDashboard')
+
+        self.timer = wpilib.Timer()
+        self.autonmodes = AutonomousModeSelector("autonomous")
 
         self.drive_controller = wpilib.XboxController(0) #0 works for sim?
 
@@ -67,6 +144,42 @@ class SpartaBot(MagicRobot):
     def disabledPeriodic(self):
         self.sd.putValue("Mode", "Disabled")
 
+    def autonomousInit(self):
+        self.timer.start()
+        self.sd.putValue("Mode", "Autonomous")
+        print("Auton Beginning")
+        # Begins the autonomous timer, should be limited at 15 seconds
+
+    def autonomousPeriodic(self):
+
+        print("Running...")
+
+        if self.timer.get() < 2.0:
+            self.drivetrain.set_motors(0.5, 0.0)
+            print(self.timer.get())
+            print("first auton state running")
+            self.sd.putValue("Drivetrain: ", "First Auton State...")
+        else:
+            self.drivetrain.set_motors(0.0, 0.0)
+            print("State 1 Finished")
+            self.sd.putValue("Drivetrain:", "First State Finished")
+
+        self.drivetrain.execute()
+
+
+        # if self.timer.get() < 3.0:
+        #     self.drivetrain.set_motors(-0.5, 0.0)
+        #     self.sd.putValue("Drivetrain: ", "Second Auton State...")
+        #     print("second state running")
+        # else:
+        #     self.drivetrain.set_motors(0.0, 0.0)
+        #     self.sd.putValue("Drivetrain: ", "Second Auton State...")
+        #     print("done")
+
+    def disabledInit(self):
+        self.autonmodes.disable()
+        self.drivetrain.set_motors(0.0, 0.0)
+        self.drivetrain.execute()
     def teleopInit(self):
         '''Called when teleop starts; optional'''
         self.sd.putValue("Mode", "Teleop")
@@ -75,7 +188,7 @@ class SpartaBot(MagicRobot):
         '''Called on each iteration of the control loop'''
 
         # drive controls
-
+        print("tele")
         angle = self.drive_controller.getRightX()
         speed = self.drive_controller.getLeftY()
 
@@ -93,53 +206,53 @@ class SpartaBot(MagicRobot):
         # if left bumper button pressed, right and left triggers control boom extension
         #   else, they control angle
 
-        if (self.drive_controller.getLeftBumper()):
-            #extend_speed = 0
-
-            # left trigger retracts, while right trigger extends
-            self.boom_arm.extender_speed -= self.drive_controller.getLeftTriggerAxis()
-            self.boom_arm.extender_speed += self.drive_controller.getRightTriggerAxis()
-
-            #self.boom_arm.set_extender(extend_speed)
-            
-        elif self.drive_controller.getRightTriggerAxis() > 0.05:
-            #rotation_speed = 0
-
-            
-            self.boom_arm.rotator_speed = self.drive_controller.getRightTriggerAxis()/10
-
-
-            #self.boom_arm.set_rotator(rotation_speed)
-            #self.boom_rotator_spark.set(rotation_speed/4)
-        elif self.drive_controller.getLeftTriggerAxis() > 0.05:
-            self.boom_arm.rotator_speed = -self.drive_controller.getLeftTriggerAxis()/10
-
-        else:
-            self.boom_arm.rotator_speed = 0
-            self.boom_arm.extender_speed = 0
-            
-        
-        if self.drive_controller.getXButton():
-            #self.testmotor.set(0.5)
-            self.boom_arm.rotator_speed = 0.5
-            print(self.boom_arm.rotator_speed)
-
-        else:
-            #self.testmotor.set(0)
-            self.boom_rotator_spark.set(0)
-
-        
+        # if (self.drive_controller.getLeftBumper()):
+        #     #extend_speed = 0
+        #
+        #     # left trigger retracts, while right trigger extends
+        #     self.boom_arm.extender_speed -= self.drive_controller.getLeftTriggerAxis()
+        #     self.boom_arm.extender_speed += self.drive_controller.getRightTriggerAxis()
+        #
+        #     #self.boom_arm.set_extender(extend_speed)
+        #
+        # elif self.drive_controller.getRightTriggerAxis() > 0.05:
+        #     #rotation_speed = 0
+        #
+        #
+        #     self.boom_arm.rotator_speed = self.drive_controller.getRightTriggerAxis()/10
+        #
+        #
+        #     #self.boom_arm.set_rotator(rotation_speed)
+        #     #self.boom_rotator_spark.set(rotation_speed/4)
+        # elif self.drive_controller.getLeftTriggerAxis() > 0.05:
+        #     self.boom_arm.rotator_speed = -self.drive_controller.getLeftTriggerAxis()/10
+        #
+        # else:
+        #     self.boom_arm.rotator_speed = 0
+        #     self.boom_arm.extender_speed = 0
+        #
+        #
+        # if self.drive_controller.getXButton():
+        #     #self.testmotor.set(0.5)
+        #     self.boom_arm.rotator_speed = 0.5
+        #     print(self.boom_arm.rotator_speed)
+        #
+        # else:
+        #     #self.testmotor.set(0)
+        #     self.boom_rotator_spark.set(0)
+        #
+        #
 
 
         # self.drivetrain's execute() method is automatically called
 
-        if self.drive_controller.getBButtonReleased():
-            grabber.turn_off_compressor(self)
-        
-        if self.drive_controller.getAButtonReleased():
-            grabber.solenoid_toggle(self)
+        # if self.drive_controller.getBButtonReleased():
+        #     grabber.turn_off_compressor(self)
+        #
+        # if self.drive_controller.getAButtonReleased():
+        #     grabber.solenoid_toggle(self)
 
-        if self.drive_controller.getYButton():
+        # if self.drive_controller.getYButton():
